+++ NEWS.md
@@ -1,6 +1,7 @@
 # mvgam 1.1.3 (development version; not yet on CRAN)
 ## New functionalities
 * Allow intercepts to be included in process models when `trend_formula` is supplied. This breaks the assumption that the process has to be zero-centred, adding flexibility but also potentially inducing nonidentifiabilities with respect to any observation model intercepts. Thoughtful priors are a must for these models
+* Added `stancode.mvgam` and `stancode.mvgam_prefit` methods
 
 ## Deprecations
 * The `drift` argument has been deprecated. It is now recommended for users to include parametric fixed effects of "time" in their respective GAM formulae to capture any expected drift effects